requirements.txt
pytelegrambotapi
telethon
python-decouple